chore(zdtp-migration): delete legacy design-token-tweak directory (W3-2)
Removes the now-orphaned src/components/design-token-tweak/ directory (21
files, 5 test files) now that the new zdtp panel is wired via
design-token-panel-bootstrap.ts + configurePanel().

- pages/lib/_body-end-islands.tsx: drop legacy DesignTokenTweakPanel Island;
  zdtp panel self-mounts as a configurePanel() side-effect, no Preact island needed
- packages/zudo-doc-v2/src/theme/design-token-tweak-panel.tsx: replace with
  null-returning stub to preserve barrel export shape during W3-2→W5-1 window
- packages/zudo-doc-v2/src/theme/color-tweak-export-modal.tsx: migrate
  ColorTweakState/TweakState imports to @/utils/design-token-types; remove
  initColorFromSchemeData (simplify resolveColorDefaults to explicit ?? fallback)
- src/config/__tests__/design-tokens-manifest.test.ts: port manifest coverage
  drift-guard tests from deleted token-manifest.test.ts and size-manifest.test.ts
- .template-drift-allowlist: add all 22 deleted paths with W3-2→W5-1 transition
  window comment
