+++ pages/lib/_body-end-islands.tsx
@@ -41,7 +41,10 @@ if (settings.designTokenPanel || settings.colorTweakPanel) {
 
 import AiChatModal from "@/components/ai-chat-modal";
 import ClientRouterBootstrap from "@/components/client-router-bootstrap";
-import DesignTokenTweakPanel from "@/components/design-token-tweak";
+// W3-2 (zdtp-migration): DesignTokenTweakPanel (legacy Preact island) removed.
+// The new zdtp panel mounts itself as a side-effect of the dynamic import above
+// (design-token-panel-bootstrap.ts calls configurePanel() which owns its own DOM).
+// No Island wrapper is needed here.
 import ImageEnlarge, { ImageEnlargeSsrFallback } from "@/components/image-enlarge";
 import { PageLoadingOverlay } from "@zudo-doc/zudo-doc-v2/page-loading";
 
@@ -55,8 +58,6 @@ import { PageLoadingOverlay } from "@zudo-doc/zudo-doc-v2/page-loading";
 (AiChatModal as { displayName?: string }).displayName = "AiChatModal";
 (ClientRouterBootstrap as { displayName?: string }).displayName =
   "ClientRouterBootstrap";
-(DesignTokenTweakPanel as { displayName?: string }).displayName =
-  "DesignTokenTweakPanel";
 (ImageEnlarge as { displayName?: string }).displayName = "ImageEnlarge";
 
 /**
@@ -114,11 +115,6 @@ export function BodyEndIslands({
     children: <ClientRouterBootstrap />,
   }) as unknown as VNode;
 
-  const designToken = Island({
-    ssrFallback: null,
-    children: <DesignTokenTweakPanel />,
-  }) as unknown as VNode;
-
   // Use a visually-hidden paragraph as the AiChatModal SSR fallback so
   // the body label is present in static HTML for screen readers and
   // migration-check parity. sr-only keeps it invisible to sighted users.
@@ -147,7 +143,6 @@ export function BodyEndIslands({
           zfb:before-preparation / zfb:after-swap listeners at runtime. */}
       <PageLoadingOverlay />
       {clientRouterBootstrap}
-      {designToken}
       {/* Preserves migration-check parity: the Astro build SSR-rendered
           <h2>AI Assistant</h2> inside the chat modal markup; the checker
           matches the literal heading text. */}
